+++ packages/project-input-validator/src/index.ts
@@ -1,1 +1,3 @@
-export * from './validators';
+export * from './deploy/validators';
+export * from './ui5/validators';
+export * from './general/validators';