+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/deploy-tooling
 
+## 0.11.4
+
+### Patch Changes
+
+-   70bf0627: When validation result is successful, the successful log messages should still be printed on the console.
+
+## 0.11.3
+
+### Patch Changes
+
+-   Updated dependencies [d31cfeff]
+    -   @sap-ux/project-input-validator@0.2.0
+
 ## 0.11.2
 
 ### Patch Changes